added filter parsing and separated calculating stats into separate util file
