+++ app/server/utils/transfer_filter.py
@@ -1,5 +1,11 @@
 import enum
 from sqlalchemy.sql import text
+from sqlalchemy import or_, Column, String, Float
+from server import db
+from server.models.custom_attribute_user_storage import CustomAttributeUserStorage
+from server.models.transfer_account import TransferAccount
+from server.models.user import User
+from sqlalchemy.sql.expression import cast
 
 MALE = 'male'
 FEMALE = 'female'
@@ -10,11 +16,6 @@ TOKEN_AGENT = 'Token Agent'
 GROUP_ACCOUNT = 'Group Account'
 ADMIN = 'Admin'
 
-USER_TABLE = 'user'
-CUSTOM_ATTRIBUTE_USER_STORAGE = 'custom_attribute_user_storage'
-TRANSFER_ACCOUNT_TABLE = 'transfer_account'
-        
-
 class TransferFilterEnum:
     INT_RANGE       = "int_range"
     DATE_RANGE      = 'date_range'
@@ -22,29 +23,34 @@ class TransferFilterEnum:
 
 TRANSFER_FILTERS = {
     'created': {
-        'table': USER_TABLE,
+        'table': User.__tablename__,
         'type' : TransferFilterEnum.DATE_RANGE,
     },
     'User Type': {
-        'table': USER_TABLE,
+        'table': User.__tablename__,
         'type': TransferFilterEnum.DISCRETE,
         'values': [BENEFICIARY, VENDOR, TOKEN_AGENT, GROUP_ACCOUNT, ADMIN]
     },
     'gender': {
-        'table': CUSTOM_ATTRIBUTE_USER_STORAGE,
+        'table': CustomAttributeUserStorage.__tablename__,
         'type': TransferFilterEnum.DISCRETE,
         'values' : [MALE, FEMALE]
     },
     'balance': {
-        'table': TRANSFER_ACCOUNT_TABLE,
+        'table': TransferAccount.__tablename__,
         'type' : TransferFilterEnum.INT_RANGE
     }
 }
 
+# will return a dictionary with table names as keys
+# values will be a 2D dictionary of tuples
+# values on the inner array should be OR'd together (mainly to account for multiple discrete values)
+# values on the outer array should be AND'd together
+
 def process_transfer_filters(encoded_filters):
     # parse and prepare filters for calculating transfer stats
     tokenized_filters = encoded_filters.split("%")
-    filters = []
+    filters = {}
     curr_keyName = None
 
     to_handle = []
@@ -55,11 +61,7 @@ def process_transfer_filters(encoded_filters):
             if symbol == ",":
 
                 # handle currently collected filters
-                if len(to_handle) > 0 and (curr_keyName is not None):
-                    if(TRANSFER_FILTERS[curr_keyName]['table'] == CUSTOM_ATTRIBUTE_USER_STORAGE):
-                        filters.append(handle_custom_user_storage_filter(curr_keyName, to_handle))
-                    else:
-                        filters.append(handle_filter(curr_keyName, to_handle))
+                filters = handle_filters_per_keyname(to_handle, curr_keyName, filters)
 
                 curr_keyName = None
                 if subject in TRANSFER_FILTERS:
@@ -72,31 +74,61 @@ def process_transfer_filters(encoded_filters):
                 })
 
     # handle currently collected filters
-    if len(to_handle) > 0 and (curr_keyName is not None):
-        if(TRANSFER_FILTERS[curr_keyName]['table'] == CUSTOM_ATTRIBUTE_USER_STORAGE):
-            filters.append(handle_custom_user_storage_filter(curr_keyName, to_handle))
-        else:
-            filters.append(handle_filter(curr_keyName, to_handle))
-
-    print(filters)
-
+    filters = handle_filters_per_keyname(to_handle, curr_keyName, filters)
     return filters
 
+def handle_filters_per_keyname(to_handle, key_name, filters):
+    if len(to_handle) > 0 and (key_name is not None):
+        curr_table = TRANSFER_FILTERS[key_name]['table']
+        if(curr_table == CustomAttributeUserStorage.__tablename__):
+            _filters = filters[curr_table] if curr_table in filters and isinstance(filters[curr_table], list) else []
+            _filters.append(handle_custom_user_storage_filter(key_name, to_handle))
+            filters[curr_table] = _filters
+        else:
+            _filters = filters[curr_table] if curr_table in filters and isinstance(filters[curr_table], list) else []
+            _filters.append(handle_filter(key_name, to_handle))
+            filters[curr_table] = _filters
+    return filters
+
+# assemble filters 
 def handle_custom_user_storage_filter(keyname, filters):
-    filter_string = ""
+    formatted_filters = []
     for i, filter_action in enumerate(filters):
         comparator = filter_action['comparator']
-        value = filter_action['value']
-        filter_string += f"{'OR' if i > 0 else ''} \"{TRANSFER_FILTERS[keyname]['table']}\".value {comparator} '\"{value}\"' "
-    print(filter_string)
-
-    return text(filter_string)
+        val = filter_action['value']
+        if comparator == '=':
+            formatted_filters.append((keyname, "EQ", val))
+        elif comparator == '>':
+            formatted_filters.append((keyname, "GT", val))
+        elif comparator == '<':
+            formatted_filters.append((keyname, "LT", val))
+        else:
+            return
+    return formatted_filters
 
 def handle_filter(keyname, filters):
-    filter_string = ""
+    formatted_filters = []
+    table = get_class_by_tablename(TRANSFER_FILTERS[keyname]['table'])
+    column = getattr(table, keyname)
     for filter_action in filters:
         comparator = filter_action['comparator']
-        value = filter_action['value']
-        filter_string += f" \"{TRANSFER_FILTERS[keyname]['table']}\".{keyname} {comparator} {value}"
-    return text(filter_string)
+        val = filter_action['value']
+        if comparator == '=':
+            formatted_filters.append((keyname, "EQ", val))
+        elif comparator == '>':
+            formatted_filters.append((keyname, "GT", val if column.type == 'DATETIME' else float(val)))
+        elif comparator == '<':
+            formatted_filters.append((keyname, "LT", val if column.type == 'DATETIME' else float(val)))
+        else:
+            return
+    return formatted_filters
 
+def get_class_by_tablename(tablename):
+  """Return class reference mapped to table.
+
+  :param tablename: String with name of table.
+  :return: Class reference or None.
+  """
+  for c in db.Model._decl_class_registry.values():
+    if hasattr(c, '__tablename__') and c.__tablename__ == tablename:
+      return c
